Move config_event to its own header
